+++ shims/fe2o3-amqp/Cargo.toml
@@ -2,6 +2,4 @@
 members = [
     "amqp_types_test",
     "amqp_large_content_test",
-    "amqp_complex_types_test",
-    "interop_test_common",
 ]
