--- shims/fe2o3-amqp/amqp_complex_types_test/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "amqp_complex_types_test"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
