--- shims/fe2o3-amqp/amqp_complex_types_test/src/bin/sender.rs
@@ -1,5 +1,0 @@
-fn main() {
-    println!("amqp_complex_types_test_data generator not implemented")
-    // let args: Vec<String> = std::env::args().collect();
-    // println!("{:?}", args);
-}
