+++ shims/fe2o3-amqp/amqp_large_content_test/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-interop_test_common = { version = "0.1.0", path = "../interop_test_common" }
 fe2o3-amqp = "0.0.31"
 anyhow = "1"
 tokio = { version = "1", features = ["net", "macros", "rt-multi-thread", "rt"]}
