--- shims/fe2o3-amqp/interop_test_common/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "interop_test_common"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-serde = "1"
-serde_json = "1"
-anyhow = "1"
-fe2o3-amqp-types = "0.0.27"
-hex = "0.4"
-ordered-float = "3"
-uuid = "1"
-base64 = "0.13"
